improvement: react build no longer includes automata.dev/core.
It expects it to be present as a dependency instead.

Also fixed the type of ActiveContent component so TypeScript understands
it better by not using the spread syntax to fish out the `children`.
This means that technically it will pass children to the
core::ActiveContent but that does not affect anything. It does mean
however that ActiveContent can never use a children.

Added DECISIONS.md so we have a place to put these decisions in to.