fix: lerna test script now works.
The link script was broken it called build before bootstrap which should
be the other way around.

The lerna test script now calls the improved link script first and then
runs all tests in each package. This way we can check for each package:
that all unit tests run, that the TypeScript compiler checks, and that
the build itself works.

Had to install `@types/estree` to get the angular package to pass
TypeScript.