feat: api docs for website.
Started working on the automata website which is for the moment very
basic. The only fleshed out part is the API generation, and even that
is still very sparse.

The pipeline works as follows: each package now uses `typedoc`, which
is a new dependency, to generate a JSON describing the package from
TypeScript and doc comments.

Each package now has a script to generate the JSON's called `apigen`,
lerna can call the apigen for each packages via `lerna run apigen`.

These JSON files with the API description are always put inside of the
website packages in a data folder. This data folder is read out by
`eleventy` which is a static site generator. From here we generate the
API part of the website using the JSON API description files.

Added and re-wrote various documentation strings for various packages.

Also rewrote code so the apigen works better:

  1. In the react package split up the hook and the component into
     separate files. This way the hook does not need to use the name
     `ActiveContentCore` but can instead use `ActiveContent` which
     makes it easier to generate links in the documentation, due to
     it not being an alias anymore.

  2. In the core package much more is now exported so it is included
     in the `typedoc` output.

  3. Various files in the vue package are now .ts files instead of .js
     files so `typedoc` can pick them up.