feat: ActiveContent now support multiple Content's at once.
Added a `maxActivationLimit` which determines how many Content's can be
active at the same time. This limit is 1 by default, when the limit is
1 it behaves in the same manner as before.

One of the biggest changes is that `active` and friends are now arrays,
to reflect that one or more Content can be active at the same time.
The `activeIndex` and `activeContent` have been renamed to
`activeIndexes` and `activeContents` to reflect this.

When you can activate multiple Content's, there must be a way to
deactivate Content's as well. So I've added methods such as
`deactivateByIndex`, `deactivate` and `deactivateByPredicate. Of course
also added a `HistoryItem` associated with deactivation. Deactivation
also works with the autoplay and cooldown.

Whenever the `maxActivationLimit` is reached what happens is determined
by the `ActiveContentMaxActivationLimitBehavior`. One of three things
can happen:

1. By default it will simply activate the Content given, and deactivate
   the first Content in the `active` array. This way a "circle" is made
   in the first in first out manner.

2. An error is thrown when the limit is reached.

3. Or nothing is done and the activation is simply ignored.

Also added the concept of the `lastActivated`, `lastActivatedIndex` and
`lastActivatedContent`, as part of this new feature. This allows the
user to quickly determine which content was the last content
which was activated.

Removed `wasActiveBeforeLast` from the `Content` as this information
lacked a clear use cases. When multiple items can be activated you
simply check the second last item in the `active` array. The reason it
was once added was to  perform a "left" animation in a carousel, but
the `isPrevious` could have been used for that instead. Due to it being
such a weird name and a thorn in my side: due to it causing many
philosophical debates, I've removed it.

Due to the `ActiveContent` now having deactivation some names caused
confusion. Take `next` and `previous` for example, what exactly do they
do? To make it really clear, all things that activate are now prefixed
by `activate`. This means that:

1. `next` is now `activateNext`

2. `previous` is now `activatePrevious`

3. `first` is now `activateFirst`

4. `last` is now `activateLast`

Another change is that `removeByPredicate` will no longer activate
the last remaining `Content` when there is only one left. This caused
the code and tests to become very complex. For the user this behavior
is also very strange, why does removing affect what becomes active? This
is still true for other methods, but I've added TODO's to remove these
as well.

Also discovered a bug in the way the `direction` was determined for when
the `isCircular` is `true`. It would not take into account that going
left or right beyond edges would be shorter. It only did so for when
the first when to the last and back, but not second going to second
last going left. Fixed this in `getDirectionWhenMovingToIndex` and
added a formula in comments on how to calculate this.