improvement: removed default activation behavior in ActiveContent.
Before when the `maxActivationLimit` was `1` on `initialization` the
first item was activated by default. This is considered unclear
behavior, and it has now been removed.

The reason it was desirable once is because the `ActiveContent` used
only allow one content to be active at a time, and then it made sense
to activate the first content by default.