improvement: removed activation behavior of first inserted element.
Before when the `maxActivationLimit` was `1` on `insertion` the
first item which was inserted was activated by default. This is
considered unclear behavior, and it has now been removed.

The reason it was desirable once is because the `ActiveContent` used
only allow one content to be active at a time, and then it made sense
to activate the inserted content by default.