improvement: removed activation behavior when active element is removed
Before when the `maxActivationLimit` was `1` on `removal` of an active
item another item was nominated to become active. This is considered
unclear behavior, and it has now been removed.

The reason it was desirable once is because the `ActiveContent` used
only allow one content to be active at a time, and then it made sense
to activate the previous element if the active element was removed.

Also renamed `ActionOptions` to `ActivationOptions` as they are now only
used for activation.