improvement: ContentPredicate argument receives more data now.
The `ItemPredicate` has been renamed to the `ContentPredicate`, and it
now receives more information than before. It used to receive the
`value` and the `index`, not it also receives the `Content` and
`ActiveContent`. This way the predicate can access more information
than before.

Also refactored out the common looping code which executes actions
based on predicates. The abstraction is called
`executeActionWhenPredicate`.