improvement: AutoplayDurationCallback argument receives more data now.
Renamed `IntervalCallback` to `AutoplayDurationCallback`, which makes
more sense. Changed all `interval` occurrences to `duration` in
documentation.

The `AutoplayDurationCallback` now receives more information than
before. It used to receive the `value` and the `index`, not it also
receives the `Content` and ActiveContent`. This way the predicate can
access more information than before.

Also changed the behavior of the `Autoplay` when there are no more
active items it stops the Autoplay. This prevents the `duration`, when
it is a callback, from ever receiving `null` as the `Content`.

Also removed `getNextIndex` and `getPreviousIndex` public methods, as
they are really an internal API. The API also makes no sense for the
general public. They have been renamed to included an underscore.

All public but private API methods should be prefixed by and underscore.
Methods with an underscore will not be included in the API generator.

Also did some minor formatting.