improvement: the cooldown now works over multiple activations.
Before the cooldown would reset for every call to activate / deactivate,
that would provide a cooldown via the activation options. This does
not make sense when supporting multiple active items thought.

Now the cooldown is more like a global "lock", which is what the user
expects.

Renamed `CooldownCallback` to `CooldownDurationCallback`, which makes
more sense.

The `CooldownDurationCallback` now receives more information than
before. It used to receive the `value` and the `index`, not it also
receives the `Content` and ActiveContent`. This way the predicate can
access more information than before.

The tests are beefed up as well, now testing mixing activation /
deactivate in combination with the cooldown. Before only activate or
deactivate where tested but not in combination with each other.