improvement: subscriber now receives event information.
Before when the subscribers were called there was no information on why
they were called. You only knew as a developer that something changed,
now we also send the `ActiveContentEvent` to every subscriber.

Renamed `HistoryItem` to `ActiveContentEvent`, this name makes it more
clear that it is no longer used just for history purposes.

Added four new events:

1. InitializedEvent: whenever the ActiveContent is initialized, this
   event is fired. It replaces the multiple activation and insertion
   events the initialize did before.

2. RemovedMultipleEvent: whenever multiple items are removed via a
   predicate this event is fired. Before it fired multiple
   RemovedEvent's

3. ActivatedMultipleEvent: whenever multiple items are activated via a
   predicate this event is fired. Before it fired multiple
   ActivatedEvent's

4. DeactivatedMultipleEvent: whenever multiple items are deactivated via
   a predicate this event is fired. Before it fired multiple
   DeactivatedEvent's

The `multiple` events were added so one action results in one history
item / event, instead of multiple events. This makes it a bit easier
to see what is happening to the ActivateContent.

Also made it possible to subscribe via the constructor, this way the
initialize event is received as well. Also had to add an unsubscribe
so the subscribe via constructor can be unsubscribed.

Renamed `shouldInformSubscribers` to `isInitializing` to make it more
clear what status it is actually tracking, updated comment to also
include history. Since history is no longer updated when initializing
due to the new `InitializedEvent`.

Also formatted the `types.ts` file.