improvement: folded in the predicate functions to save space.
Before methods which used predicates would have three versions, one
for `before` one for `after` and one for `at`. These versions would
determine where to insert / move the item to. It basically acted as a
sort of modifier.

But having all these "methods" takes up a lot of space. Both in the
surface of the API, and in literal bytes. To reduce this all such
methods now accept a `mode` in an `option` object, which is either
"before", "after" or "add".

Went from 16.699 bytes to about 16.058 bytes.