improvement: renamed ActiveContent to ActiveList, and clarified errors.
Added document comments to all errors in the `ActiveList`, also in tests
we now check if the errors are an instance of these classes.

Discovered an edge case whilst doing this. When the cooldown error
`ActiveListCooldownDurationError` occurs when the cooldown comes from
the `ActiveListActivationOptions`, it will have activated / deactivated
the item already. This is now documented as a warning since there is not
much we can do about this. The cooldown must happen after activation /
deactivation, otherwise we cannot pass the last active.

Errors now have a custom type in the documentation website.

Also prefixed all types and classes of the `ActiveList` functionality
with `ActiveList`. This makes naming collision in the future impossible.

Also some minor improvements:

  1. Added an index barrel to the `ActiveContent` folder so the
     `src/index` can re-export from there, in the future this will make
     the `src/index` leaner when multiple classes are added.

  2. Moved `UnsubscribeFunction` type to a generic folder, since it will
     be used by other classes as well.

  3. In the documentation website a hook accidentally had a `$` appended
     to its icon, it has now been removed.

  4. Added explicit `typescript` dependency to all packages.

  5. Changed the lerna package.json test command: the website is now
     built as the last step. It used to be build via the "build" command
     of the website package, but this does not work, because the API had
     not been generated yet. This causes a circular dependency. To fix
     this the last two steps now call the apigen, and then a `build-docs`
     command which only the website supports. Website now longer does
     anything on `build` now. This breaks the circular dependency, and
     all is well now.