improvement: ActiveList now longer requires a config.
Before the `ActiveList` required a config with at least the `contents`
property set. This was annoying as it required the user to have content,
but why can a list not start with any content?

The question can only be answered with: no it is not required and it
can be empty, because things can be added afterwards.