fix: the angular package published wrong directory.
The `ng-packagr` library bundles everything up into the "dist" folder,
even generating a `package.json` file there. It does this because it
expects publishing to happen from the "dist" folder.

Unfortunately lerna cannot handle this situation, the solution is to
simply publish normally. This means that we now ignore `ng-packagr`
dist/package.json, and publish the `dist` folder instead.

The `before-publish` also pulls in the LICENSE.md and README.MD files in
from the root into the `dist` folders. This way the package looks
nicer on NPM. In the angular package it also removes the
dist/package.json.

Also now running `test` on publish so every test is ran, and also the
builds are ran. This should prevent nasty surprises.