feat: digital ocean static doc site deployment.
Altered the `build-docs` so that it first links all dependencies
via `link`, then generates the api files first via `apigen` and
then runs the `build-docs`. This way one command can be executed
to build the docs in a CI environment.

Changed `nx` cache director to .cache,  so it does not create a cache
in the node_modules, which digital ocean does not allow.