fix: autoplay double pause bug.
When the autoplay is paused multiple times in a row, without unpausing
between them a bug could occur which caused a negative duration, when
it was finally unpaused.

This code here calculates the remaining duration after a pause:

```javascript
if (this.pauseStarted) {
  return (
    this.autoplayCurrentDuration -
    (this.pauseStarted.getTime() - this.autoplayStarted.getTime())
  );
}
```

It uses the `this.pauseStarted` to get the time of the pause. But what
happens if the user called `pause` in the old code:

```javascript
public pause(): void {
  // Store the time when the pause was pressed, so we can calculate
  // the resuming duration later.
  this.pauseStarted = new Date();

  this.cancelTimer();
}
```

It simply overrides the `pauseStarted`.

So when the user writes code which calls `pause` multiple times by
accident, the `pauseStarted` date is updated / mutated. This is
unexpected because `pauseStarted` should be inside the original timebox
the `duration` it paused was in.

By moving `pauseStarted` (by calling pause) outside of the original
duration a negative number could be returned.

The fix was to simply ignore pause whenever the `pauseStarted` is
defined.