fix: typings now include comments
Without comments there are no doc comments as well. Without doc
comments autocomplete in editors miss this information in their
intellisense popups.

Fixed by adding setting `removeComments` to `false` for the React, and
Vue packages. The Angular package was already properly configured.

The `core` package needed more work, because is contains massive
comments only needed for development. The fix is generating the build
twice, once with comments and once without. This way the `code` does
not include any comments, but the typings do.