fix: API json data is now always generated when running 'npm run apigen'
This was broken due to `nx` caching the apigen command, but not
understanding where the output actually is, which is in
`website/src/_data/api`.

This caused the directory to be empty on cleanup, and nx considering
this "OK" due to the cache.

By simply not caching `apigen` anymore this issue is resolved.