fix: ActiveListContent's swapWith- and Next could go out of bounds.
This happened when calling `swapWithNext` on the last item, and by
calling `swapWithPrevious` on the first item.

The reason for this is because both called the `_getUnbounded` variants
to get the swap to index. The `_getUnbounded` variants do not guard
against going out of bounds, as their names imply.

To fix this both `swapWithNext` and `swapWithPrevious` use the
`_getBounded` variants. The variants had to be altered to support
providing an index, instead of always using the `lastActivatedIndex`
of the `ActiveList`, all previous call sites now provide the
`lastActivatedIndex`, so they call the old behavior.

Also now testing all scenario's when using `swapWithNext` and
`swapWithPrevious`, on both a circular and non circular `ActiveList`.

Added more comments as to what the difference is between the two
`_getUnbounded` and `_getBounded` variants. So picking the right one in
the future is easier, to prevent this kind of mistake.