feat: minification per component.
Made it possible to minify the core on a per component basis. This means
that for each component of "uiloos" we generate a `min.js` file. This
way the user can pick the components they want to use more granularly.

To make this possible the license checker had to become a class
singleton, instead of a module. This way it became possible to separate
the ActiveList (a component) from the checking mechanism. This is
because `terser` understands the intention of a class better, and leaves
it as a complete unit. When it still was a function terser would rip out
if conditionals inside of the functions, causing the `checkLicense` to
become the faulty:

```js
// Tersers attempt at minification goes terribly wrong:
_checkLicense() {
  console.log('Warning here...')
}
```

In order to get the license checker separated from component the import
`import * as uiloosLicenseChecker from '../license/license'` has to be
removed before minifying, otherwise `terser` will inline it. This was
solved by creating a `rollup` plugin which comments this line in and
out.

To make the `rollup` config easier to understand it is now  separated
into three files: one for the code, one for the typings generation,
and one for the minification.

Also the minification itself is now more aggressive, it minifies
everything starting with an underscore. Prefixed all private fields
variables, and helper classes with underscores for this reason.

The minification's output is smoke tested in the file called:
`minification-check.js`. The test does some minor assertions to see
the function.

Also fixed a minor bug the `publish` phase: the name of the minified
files included the old version of the package from before the release.
This caused `0.0.3` minification to be called `0.0.2`, which is
incorrect. Fixed by setting the version in the `prepublishOnly` npm
lifecycle script, just before the package is uploaded.