improvement: ActiveList can now be toggled.
This allows for toggling like the `ActiveListContent` already supported.
There are two new methods on the `ActiveList` the first is `toggle`
which toggles by identity, and the second is `toggleByIndex`, which
toggles based on the index.

ActiveListContent's toggle is now re-written to use the ActiveList's
`toggleByIndex`.

Also added a unit test for the license which tests the index exports.