improvement: README now includes useful links.
This will make the NPM pages for the packages nicer to read. In order
for this to work the README's need to be put into the root
of the package and not in the dist as previously was the case. Fixed
by not copying the README in the `before-publish` to the dist but
to the root.

Also removed angular's readme which it accidentally had.