improvement: added publish:alpha script, improved test script perf
Accidentally released a `0.0.6-alpha.0` version to latest because of
a missing `-dist-tag`, by adding this new `publish:alpha` we can avoid
this in the future.

The `test` script runs `run build-docs` as the final command which
runs the `apigen` script, this makes `test` run of `apigen` unnecessary,
so it has been removed.