improvement: ActiveList cooldown has events, improved isUserInteraction
The cooldown functionality of the ActiveList now has events that tell
the user that the cooldown was started and ended. This allows for
disabled states to be visually shown.

Technically this meant rewriting the CooldownTimer to use a setTimeout
instead of checking for deltas on dates. This makes is a little less
accurate, but it is accurate enough. This is also the way AutoPlay
works already.

Also changed the behavior of the `isUserInteraction` when `undefined`
before it would treat `undefined` as `false`, now it will treat
`undefined` as `true`. The reason for this clear from the following
snippet:

```js
activeList.activateByIndex(0, { cooldown: 1000 });
```

The users intent is activating index 0 with a cooldown of `1000`, but
here the `isUserInteraction` is `undefined` so it would treat it as not
being a user interaction, which would bypass the cooldown. This is
not the expected behavior at all.

Also the same rings true for the AutoPlay, the AutoPlay should stop
when `stopsOnUserInteraction` is `true`, the snippet above would not
have stopped it.

Finally changed some of the tests for AutoPlay, added more describes.