fix: apigen execution order caused it to fail.
Due to an update of nx and lerna the order in which the apigen was
executed changed. Before the `website` always ran first, in which it
cleared the "data" folder, after the update it ran third. This caused it
to clear the data folder at half way point, which is of course not a
good idea.

The fix was adding a `apigen:pre` task to the `website`, and running it
before the `apigen` in the root package. This way we guarantee that
`website` is run first.