feat: the ActiveList now knows whether or not it has been stopped before
Use case: say you are making a carousel which should stop on any user
interaction, so you set `stopsOnUserInteraction` to `true`. Say you also
have another feature: a pause whenever the user hovers over the
carousel. These two features would cause a conflict:

Whenever the mouse over happens you call `play()`, which negates the
`stopsOnUserInteraction`, causing the carousel to play again.

To fix this problem you should on the mouse over not call `play()`
whenever `hasBeenStoppedBefore` is `true`.

The solution was adding a new variable to the `autoPlay` object called
`hasBeenStoppedBefore` this allow the developer to check if the autoPlay
has ever stopped before, and allowing the `play()` on the mouse over to
be called conditionally.

The carousel in on the home page on the `website` now uses
`hasBeenStoppedBefore` to fix this conflict.

The `website` now depends on uiloos `^1.0.0`.

Also fixed a bug, the `initialize` method of the `ActiveList`, did not
reset the `autoPlay and `cooldown` objects.

Also made the `autoPlay and `cooldown` objects of the `ActiveList`
`readonly` so they cannot be reassigned on accident. Same was done for
the `autoDismiss` on the `ViewChannelView`.