improvement: added a releases page to website and @since to all code.
On the releases pages all releases of "uiloos" are documented. It shows
the date of the release and the changelog of the release.

All methods / classes and functions now have in their docstrings a
@since comment. This is then used to link from the documentation to a
specific release. Had to rewrite a lot if "website" includes to now
be macros to get this to work.

Also in the main "package.json" added a "updated" script, it will show
which packages will need to be updated, if a release were to be made
at this moment.

Removed the "method.njk" include as it was not used anywhere.