improvement: website api docs comments are rendered with better breaks.
Before we used the built in `nl2br` filter but this filter adds line
breaks for each `\n` which does not look nice, as it breaks quite
randomly for the reader

Now we add a double `<br \>` for each `\n\n` by using a custom filter,
which makes it look more as the doc comment intends. Added a macro
called `displayComment` so every comment uses the same `code`.

Tiny doc fix: in the file `typealias.njk` the `sinceRoot` was misspelled
in the call to `displaySignatures`. This caused the link to be broken.