website: fixed search result url going to wrong API url.
For example the following url was created for ActiveList::push:

`https://www.uiloos.dev/api/core/ActiveList/#push`

This does not work on production as it is case sensitive there, it should
have been:

`https://www.uiloos.dev/api/core/active-list/#push`.

The fix is lowercasing the generated `link` so it works in production.