website: added new pages: overview, TypeScript, rules an future
In the overview page we describe uiloos in a high over fashion.
Including when it is useful, how to use it, and with an overview of all
available components.

In the TypeScript page we describe the full TypeScript support uiloos
offers.

In the rules page the rules from the ActiveList and ViewChannel have
been merged together.

In the future page we put the upcoming features of uiloos.

The new overview page is put in the getting started section, the other
pages are in the new miscellaneous section.

Also improved the getting started HUB page, changed the text added link
to activation page.

Minor changes:

  1. Fixed spelling mistakes in several pages.

  2. Removed h2 class since it is not used from various pages.

  3. Added mb-4 instead of mb-2 to h2 style.

  4. The .high CSS class is no longer limited to span's and a elements.