improvement: fixed auto toc highlighter.
No longer using a `IntersectionObserver` but instead a simple scroll
listener and an algorithm to find the closest h2 to the user. This
approach is much more stable and does not skip links like the old
implementation did.

Also fixed the headings of the getting started overview page, they were
not correctly numbered.