website: command-k search links to methods / properties were broken.
The reason for this is because the links where lowercased in the
command-k search, but not in the API pages themselves.

Fixed by making all API links lowercased by default to prevent this.

Also closing the search modal whenever the user clicks on a result,
this prevents the modal from staying visible if the user clicks on
a result he / she is already on.