bugfix: the ActiveList could not be paused and resumed again
The bug was that the variable `this._pauseStarted` in the `AutoPlay` was
not reset to `null` when play resumed. This caused all subsequent
durations to be calculated using the last irrelevant
`this._pauseStarted`.

This could then trigger a `ActiveListAutoPlayDurationError` if the
duration became negative.

Added a test in for the `ActiveList` and `ViewChannel` which tests if
they can be paused and continued multiple times.

For the ViewChannel's `AutoDismiss` there is no bug, because
`AutoDismiss` is not shared between views. I've added a test anyway
just to make sure, and added a comment explaining why it differs from
the `AutoPlay`.