feature: `ActiveList` now knows which item was the last deactivated item
Tracked in three new variables:

A. lastDeactivated: which value was the last value to be deactivated.

B. lastDeactivatedContent: which ActiveListContent was the last to be
   deactivated.

C.lastDeactivatedIndex: the index which was last deactivated.

Also discovered a bug: `activateByPredicate` does specify how to handle
`maxActivationLimit` correctly. It is not specified what should happen
when there is a limit greater than 1, and what happens when the
predicate matches more items than the `maxActivationLimit`.

This is a specification failure, it will be fixed in the next commit.