fix: ActiveListActivatedEvent did not say which items were deactivated
The `activateByIndex` emits an `ActiveListActivatedEvent` but did not
include information about which items were deactivated.

A deactivation can occur when the ActiveList's `maxActivationLimit`
is set to a number, and the `maxActivationLimitBehavior` is set to
"circular". In this scenario whenever more items are activated than
alloted, the first item that was activated gets deactivated, in a
first in first out manner.

To fix two new properties are added to the `ActiveListActivatedEvent`:
`deactivatedValue` and `deactivatedIndex`.

To detect if anything was deactivated we store what the last deactivated
index was  `activateByIndex` was called and compare them with what it
was after it is finished.