feat: added new concept for vanilla js / ts usage: method subscribers.
A method subscriber is a function that takes a component and listens
to each event, and then calls a method which is provided.

For example whenever the `ACTIVATED` event occurs it will call the
`onActivated` method. This way the user doe not have to write as many
if-statements and switches in order to distinguish between events.

Had to make changes to the website to support these new functions:

1. Examples can now be made full screen when they are given the
   `w-full` CSS class. Needed since 1.5.0 release notes shows a code
   example which looked better full screen.

2. `sortApi` now places classes above functions.

3. `distil-api.js` for type definitions now sets the type of a property
   which is a function to be of type function.