feat: added opposite direction to ActiveList.
Also fixed a bug in the deactivation reset routine: whenever the
ActiveList is in a completely inactive state due to a deactivation,
the direction was reset to a hard-coded "right". Which is incorrect
because the directions are configured in the config. Now it is set
to whatever direction was configured for `next`.

Added tests to verify that the bug was fixed, improved the test for
a custom direction so it no longer uses a "call times check" but
actually checks the complete state.