docs: ActiveList improved maxActivationLimit documentation
Fixed wording in ActiveList:

1. the `maxActivationLimitBehavior` mode 'ignore'.
2. there are the values ->  these are the values
3. added explanation on why an activation can cause a deactivation
   in `ActiveListActivatedMultipleEvent`.
4. In `activateByPredicate` expanded on what happens when mode
   is "error" when the activation limit is reached.

ActiveList docs fixed non closing backticks around the word null.

Also the `constructor` and `initialize` can in fact throw a
`ActiveListActivationLimitReachedError`. This was always the case but
they were not documented in the doc comments. Added a test to prove
that this behavior does indeed occur.