fix: TypeScript types for Vue did not generate correctly.
Problem was that the types were not created correctly and set in the
package.json.

To fix the type generation the package `vite-plugin-dts` was added.
The types are not generated on build in the vite.config.js via the
new package.

As for the package.json, its type set to "module" now, and the types
are now included in the "export" key.